refactor: TransactionManager 분리
