+++ jdbc/src/main/java/com/interface21/jdbc/datasource/DataSourceUtils.java
@@ -3,7 +3,6 @@ package com.interface21.jdbc.datasource;
 import com.interface21.jdbc.CannotGetJdbcConnectionException;
 import com.interface21.transaction.support.TransactionSynchronizationManager;
 import java.sql.Connection;
-import java.sql.SQLException;
 import javax.sql.DataSource;
 
 // 4단계 미션에서 사용할 것
@@ -13,14 +12,5 @@ public abstract class DataSourceUtils {
 
     public static Connection getConnection(DataSource dataSource) throws CannotGetJdbcConnectionException {
         return TransactionSynchronizationManager.getResource(dataSource);
-    }
-
-    public static void releaseConnection(DataSource dataSource) {
-        try {
-            Connection connection = TransactionSynchronizationManager.unbindResource(dataSource);
-            connection.close();
-        } catch (SQLException ex) {
-            throw new CannotGetJdbcConnectionException("Failed to close JDBC Connection");
-        }
     }
 }
